One line OK's and fix functions for compile libft.a

+++ libft-WIP/srcs/ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/22 22:38:24 by sbednar           #+#    #+#             */
-/*   Updated: 2018/11/22 23:37:19 by sbednar          ###   ########.fr       */
+/*   Updated: 2018/11/23 00:22:21 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	ft_bzero(void *s, size_t n)
 {
 	char	*p;
-	int		i;
+	size_t	i;
 
 	p = (char *)s;
 	i = 0;
