+++ libft-WIP/srcs/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/22 22:56:15 by sbednar           #+#    #+#             */
-/*   Updated: 2018/11/22 23:37:06 by sbednar          ###   ########.fr       */
+/*   Updated: 2018/11/23 00:22:36 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memcpy(void *dst, void const *src, size_t n)
 {
 	char	*csrc;
 	char	*cdst;
-	int		i;
+	size_t	i;
 
 	csrc = (char *)src;
 	cdst = (char *)dst;
